cover image should be to the left and intro should be to the right in desktop
