+++ app.js
@@ -1,7 +1,7 @@
 
 import './js/effects.js';
-import './js/aside_template.js';
 import './js/prism.js';
+import './js/submenu_template.js'
 import './js/about_template.js';
 import './js/skills_template.js';
 import './js/testimonials_template.js';
