+++ index.html
@@ -34,29 +34,32 @@
         <a class="to_nav smooth" href="#primary_nav">menu <i class="fa fa-angle-down"></i></a>
       </header>
       <main>
-        <section id="mobile-cover" class="cover">
+        <section id="cover-img">
           <img src="img/cover-img.svg" alt="profile">
           <h1>I am a Front End &<br/>Ruby on Rails Developer</h1>
-          <div id="submenu">
-            <a href="#testimonials" class="smooth">testimonials <i class="fa fa-quote-right"></i></a>
-            <a href="#featured-code" class="smooth">featured code <i class="fa fa-code"></i></a>
-            <a href="mailto:alejandro.ginesmartinez@gmail.com?subject=Contact%20from%20alexgines.com." rolle="button">contact me <i class="fa fa-envelope-o"></i></a>
+          <div id="submenu-mobile">
+          <!-- insert here submenu for mobile version -->
           </div>
         </section>
-        <section id="typewriter1">
-          <h1>
-            <a href="" class="typewrite" data-period="3000" data-type='[ "I focus on clean, readable code.", "I am determined to excel.", "I thrive on daily challenge.", "I’m always willing to go the extra mile." ]'>
-              <span class="wrap"></span>
-            </a>
-          </h1>
-        </section>
-        <section id="intro">
-          <ul>
-            <li>Put simply, coding is my passion. And I’m proud of making my passion into a living.</li>
-          </ul>
+        <section id ="cover">
+          <div id="typewriter1">
+            <h1>
+              <a href="" class="typewrite" data-period="3000" data-type='[ "I focus on clean, readable code.", "I am determined to excel.", "I thrive on daily challenge.", "I’m always willing to go the extra mile." ]'>
+                <span class="wrap"></span>
+              </a>
+            </h1>
+          </div>
+          <div id="intro">
+            <ul>
+              <li>Put simply, coding is my passion. And I’m proud of making my passion into a living.</li>
+            </ul>
 
+          </div>
         </section>
       </main>
+      <div id="submenu-desktop">
+        <!-- insert here submenu for desktop version -->
+      </div>
       <section id="about">
         <ul class="section-heading">
           <li>
@@ -142,13 +145,10 @@
           <li><a href="#submenu" class="smooth">About</a></li>
           <li><a href="#featured-code" class="smooth">Code</a></li>
           <li><a href="http://tutorialzine.com/2014/09/50-awesome-tools-and-resources-for-web-developers/" target="blank">Resources</a></li>
-          <li><a href="http://alexgines.com/AlejandroGinesMartinezCV.pdf" id="resume" type="application/pdf">Resume</a></li>
           <li><a href="mailto:alejandro.ginesmartinez@gmail.com?subject=Contact%20from%20alexgines.com." rolle="button">Contact</a></li>
+          <li><a href="http://alexgines.com/AlejandroGinesMartinezCV.pdf" id="resume" type="application/pdf">Resume</a></li>
         </ul>
       </nav><!-- end primary-nav-->
-      <aside id="cover-sidebar" class="cover">
-      <!-- here goes aside_template -->
-      </aside>
     </div><!--end wrapper-->
 
     <!-- script tags with JavaScript files here -->
