+++ prepros (Alejandro Ginés Martínez's conflicted copy 2017-05-10).cfg
@@ -382,54 +382,6 @@
           "imported": false,
           "parents": [],
           "type": "javascript"
-        },
-        "be79ad81": {
-          "id": "be79ad81",
-          "path": "bundle (Alejandro Ginés Martínez's conflicted copy 2017-05-10).js",
-          "output": "bundle (Alejandro Ginés Martínez's conflicted copy 2017-05-10)-dist.js",
-          "name": "bundle (Alejandro Ginés Martínez's conflicted copy 2017-05-10).js",
-          "category": "JS",
-          "autoCompile": false,
-          "uglify": true,
-          "mangle": true,
-          "babel": false,
-          "sourceMaps": false,
-          "customOutput": false,
-          "imported": false,
-          "parents": [],
-          "type": "javascript"
-        },
-        "892a7c9d": {
-          "id": "892a7c9d",
-          "path": "css/main (Alejandro Ginés Martínez's conflicted copy 2017-05-10).css",
-          "output": "css/main (Alejandro Ginés Martínez's conflicted copy 2017-05-10)-dist.css",
-          "name": "main (Alejandro Ginés Martínez's conflicted copy 2017-05-10).css",
-          "category": "CSS",
-          "autoCompile": false,
-          "sourceMaps": false,
-          "compress": true,
-          "cssnext": false,
-          "autoprefixer": false,
-          "customOutput": false,
-          "imported": false,
-          "parents": [],
-          "type": "css"
-        },
-        "7f0800fa": {
-          "id": "7f0800fa",
-          "path": "js/submenu_template.js",
-          "output": "js/submenu_template-dist.js",
-          "name": "submenu_template.js",
-          "category": "JS",
-          "autoCompile": false,
-          "uglify": true,
-          "mangle": true,
-          "babel": false,
-          "sourceMaps": false,
-          "customOutput": false,
-          "imported": false,
-          "parents": [],
-          "type": "javascript"
         }
       },
       "deploymentHistory": {},
